ClickOH/requirements.txt
django==4.0.2
djangorestframework==3.13.0
requests[security] == 2.9.1